scripts/requirements.txt
# Requirements para geração de vídeos com Grok
requests>=2.31.0
python-dotenv>=1.0.0
